COM-2013 - change score from activities to cards store
