+++ src/screens/courses/cardTemplates/OrderTheSequenceCard/index.tsx
@@ -7,7 +7,7 @@ import { useNavigation } from '@react-navigation/native';
 import { footerColorsType, OrderedAnswerType, OrderTheSequenceType } from '../../../../types/CardType';
 import { StateType } from '../../../../types/store/StoreType';
 import Selectors from '../../../../store/cards/selectors';
-import Actions from '../../../../store/activities/actions';
+import Actions from '../../../../store/cards/actions';
 import CardHeader from '../../../../components/cards/CardHeader';
 import { GREEN, GREY, ORANGE, PINK } from '../../../../styles/colors';
 import QuizCardFooter from '../../../../components/cards/QuizCardFooter';
