+++ src/screens/courses/cardTemplates/SingleChoiceQuestionCard/index.tsx
@@ -5,7 +5,7 @@ import shuffle from 'lodash/shuffle';
 import { footerColorsType, SingleChoiceQuestionType } from '../../../../types/CardType';
 import { StateType } from '../../../../types/store/StoreType';
 import Selectors from '../../../../store/cards/selectors';
-import Actions from '../../../../store/activities/actions';
+import Actions from '../../../../store/cards/actions';
 import CardHeader from '../../../../components/cards/CardHeader';
 import { GREY, GREEN, ORANGE, PINK } from '../../../../styles/colors';
 import QuizCardFooter from '../../../../components/cards/QuizCardFooter';
