+++ src/store/activities/reducers.ts
@@ -3,7 +3,6 @@ import {
   SET_ACTIVITY,
   SET_EXIT_CONFIRMATION_MODAL,
   RESET_ACTIVITY_REDUCER,
-  INC_GOOD_ANSWERS_COUNT,
   SET_ACTIVITY_HISTORIES,
   ActivityActionType,
   ActivityActionWithoutPayloadType,
@@ -12,7 +11,6 @@ import {
 const initialState: ActivityStateType = {
   activity: null,
   exitConfirmationModal: false,
-  score: 0,
   activityHistories: [],
 };
 
@@ -29,8 +27,6 @@ export const activities = (
       return { ...state, activityHistories: action.payload };
     case RESET_ACTIVITY_REDUCER:
       return initialState;
-    case INC_GOOD_ANSWERS_COUNT:
-      return { ...state, score: state.score + 1 };
     default:
       return state;
   }
